requirements.txt
streamlit
numpy
Pillow
opencv-python
pandas
numpy
seaborn
matplotlib
requests

